+++ src/Credfeto.Aur.Mirror.Server/Program.cs
@@ -9,6 +9,7 @@ using Credfeto.Aur.Mirror.Server.Helpers;
 using Credfeto.Aur.Mirror.Server.Middleware;
 using Credfeto.Docker.HealthCheck.Http.Client;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.Logging.Abstractions;
 
 namespace Credfeto.Aur.Mirror.Server;
 
@@ -24,7 +25,11 @@ public static class Program
     public static async Task<int> Main(string[] args)
     {
         return HealthCheckClient.IsHealthCheck(args: args, out string? checkUrl)
-            ? await HealthCheckClient.ExecuteAsync(targetUrl: checkUrl, cancellationToken: CancellationToken.None)
+            ? await HealthCheckClient.ExecuteAsync(
+                targetUrl: checkUrl,
+                logger: NullLogger.Instance,
+                cancellationToken: CancellationToken.None
+            )
             : await RunServerAsync(args);
     }